+++ js/functions.js
@@ -1,17 +1,11 @@
-// Функция для проверки длины строки.
-//Не беспокоился о типах данных аргументов и о том, как должна вести себя функция, если ей были переданы неверные аргументы.
 function getMyStr (str, mynumber) {
-
   return str.length <= mynumber;
 }
 
 getMyStr ('длиннаястрока', 10);
 getMyStr ('длиннаястрока', 30);
 
-//Функция для проверки, является ли строка палиндромом, вариант как по заданию!
-
 function getStrgPol (string) {
-
   let strNorm = string.replaceAll(' ', '');
   strNorm = strNorm.toUpperCase();
   let strReverse = '';
@@ -19,27 +13,26 @@ function getStrgPol (string) {
   for (let i = strNorm.length - 1; i >= 0; i--) {
     strReverse += strNorm[i];
   }
-
   return strNorm === strReverse;
-
 }
 getStrgPol ('дед');
 getStrgPol ('Шалаш');
 getStrgPol ('Шалаш Шалаш Шалаш');
 getStrgPol ('Интересное задание спасибо за проверку');
 
-
-//Функция для проверки, является ли строка палиндромом - мой вариант!
-/*
-function checkStringPolidrom  (str) {
-
-  let string = str.replaceAll(' ', '').toUpperCase();
-  for (let i = 0; i < string.length; i++) {
-      if(string[i] === string[(string.length-1) - i]){
-          return true;
-      }
-      return false;
-  }
+function checkNormLengthDay (dayStart,dayEnd, startConference, lengthConference) {
+  const dayStartSplit = dayStart.split(':');
+  const dayStartMin = Number(dayStartSplit[0]) * 60 + Number(dayStartSplit[1]);
+  const dayEndSplit = dayEnd.split(':');
+  const dayEndMin = Number(dayEndSplit[0]) * 60 + Number(dayEndSplit[1]);
+  const startConferenceSplit = startConference.split(':');
+  const startConfernceMin = Number(startConferenceSplit[0]) * 60 + Number(startConferenceSplit[1]);
+  const endConferenceMin = startConfernceMin + Number(lengthConference);
+  return dayStartMin <= startConfernceMin && endConferenceMin <= dayEndMin;
 }
-*/
 
+checkNormLengthDay('08:00', '17:30', '14:00', 90);
+checkNormLengthDay('8:0', '10:0', '8:0', 120);
+checkNormLengthDay('08:00', '14:30', '14:00', 90);
+checkNormLengthDay('14:00', '17:30', '08:00', 90);
+checkNormLengthDay('8:00', '17:30', '08:00', 900);
