+++ js/main.js
@@ -1,4 +1,2 @@
-import {getPhotos} from './create-photo-arrays.js';
-const MAX_LENGTH_PHOTOS = 25;
+import './photos/thumbnails.js';
 
-Array.from({length:MAX_LENGTH_PHOTOS}, getPhotos());
